shallot: move the native host to rust/native, split prebuilt fetch
