+++ rust/native/icon.ico
