+++ rust/native/icon.rc
