+++ rust/native/src/helper.rs
