+++ scripts/check-pack.ts
@@ -54,8 +54,8 @@ const required = [
     "src/cli/index.ts",
     "dist/vite.js",
     "dist/harness-browser.js",
-    "rust/window/Cargo.toml",
-    "rust/window/Cargo.lock",
+    "rust/native/Cargo.toml",
+    "rust/native/Cargo.lock",
     "assets/icon-1024.png",
     "AGENTS.md",
     "MIGRATION.md",
